:technologist: Add story for profile button
